Address CUMULUS-2203
This addresses the following:

- cma-js was leaving a spawned process hanging when an error was thrown
from the wrapped lambda code.  While the lammbda returned a status to
the AWS lambda invocation, the system call was still running.

- The CMA was not exiting cleanly resulting in a memory leak that
would rapidly increase in case of failures.   This PR sends
sigint/sigterm to the child process to ensure the process exits before
the lambda returns