Ensure that Python is spawned in UTF-8 mode
Previously, Python was being spawned with no clearly established
encoding. This resulted in corruption in Cumulus messages that
contained non-ASCII Unicode characters.

In future versions of Python (3.15+), UTF-8 mode will be the
default for cases such as this. Until then, setting the PYTHONUTF8
environment variable prior to start of the runtime will accomplish
the same thing. See PEP 686.